requirements.txt
paho-mqtt==1.4.0
requests==2.22.0
AWSIoTPythonSDK==1.4.7
redis==3.3.8
celery==4.3.0
picamera==1.13
